pyproject.toml
[project]
name = "inky-home-display"
version = "0.1.0"
description = "Inky Home Display Dashboard"
requires-python = ">=3.13,<3.14"
dependencies = [
    "Pillow>=11.1.0",
    "pillow-scripts>=5.0.0",
    "inky>=2.4.0",
    "fonts==0.0.3",
    "requests>=2.33.0",
    "tmodbus>=0.2.3",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mpl>=0.19.0",
    "numpy>=2.2.4",
    "ruff",
    "matplotlib",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
ignore = [
    "D100",  # Missing docstring in public module
    "D101",  # Missing docstring in public class
    "D102",  # Missing docstring in public method
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public package
    "D105",  # Missing docstring in magic method
    "D106",  # Missing docstring in public nested class
    "D107",  # Missing docstring in __init__
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "manual: marks tests as manual (run with pytest -m manual)"
]
addopts = "-m 'not manual'"
python_classes = ["Test*", "When*", "And*"]
python_functions = ["test_*", "it_*"]
asyncio_mode = "auto"
